+++ core/control-plane/control-plane-transform/build.gradle.kts
@@ -22,6 +22,7 @@ dependencies {
     api(project(":spi:common:participant-spi"))
     api(project(":spi:common:transform-spi"))
     api(project(":spi:control-plane:asset-spi"))
+    api(project(":spi:control-plane:policy-spi"))
 
     testImplementation(project(":tests:junit-base"));
     testImplementation(project(":core:common:lib:json-ld-lib"))
