+++ extensions/control-plane/api/management-api/policy-definition-api/build.gradle.kts
@@ -29,6 +29,8 @@ dependencies {
     implementation(project(":core:common:lib:api-lib"))
     implementation(project(":core:common:lib:util-lib"))
     implementation(project(":core:common:lib:validator-lib"))
+    implementation(project(":core:control-plane:control-plane-transform"))
+
     implementation(project(":extensions:common:api:lib:management-api-lib"))
     implementation(project(":extensions:common:http:lib:jersey-providers-lib"))
